requirements.txt
streamlit==1.32.0
scikit-learn==1.3.0
python-docx==0.8.12
nltk==3.8.1
